# niakofa-repo/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Niakofa - Map-First Community Help Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create a business application (starts pending admin approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessInput"
      responses:
        "201":
          description: Business created (pending approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
...
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get a business (members only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "403":
          description: Not a member
        "404":
          description: Not found
    patch:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update business details (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessInput"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}/members:
    get:
      operationId: getBusinessMembers
      tags: [businesses]
      summary: List all members of a business (members only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessMember"
    post:
      operationId: inviteBusinessMember
      tags: [businesses]
      summary: Invite a user by email (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessMemberInviteInput"
      responses:
        "201":
          description: Member invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessMember"
  /businesses/{id}/members/{memberId}:
    delete:
      operationId: removeBusinessMember
      tags: [businesses]
      summary: Remove a member (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member removed
  /businesses/{id}/requests:
    get:
      operationId: getBusinessRequests
      tags: [businesses]
      summary: List all requests posted under a business (members only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of business requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessRequest"
  /businesses/{id}/members/{memberId}/cap:
    patch:
      operationId: setBusinessMemberCap
      tags: [businesses]
      summary: Set a staff member's spending cap (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessMemberCapInput"
      responses:
        "200":
          description: Updated member with cap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessMember"
  /businesses/{id}/pending-requests:
    get:
      operationId: getBusinessPendingRequests
      tags: [businesses]
      summary: List staff requests pending owner approval (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of pending requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessPendingRequest"
  /businesses/{id}/requests/{requestId}/approve:
    patch:
      operationId: approveBusinessRequest
      tags: [businesses]
      summary: Approve or reject a pending business request (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessRequestApproveAction"
      responses:
        "200":
          description: Request approved or rejected
...
components:
  schemas:
...
    Business:
      type: object
      required: [id, legal_name, display_name, approval_status, created_by_user_id, created_at, updated_at]
      properties:
        id:
          type: integer
        legal_name:
          type: string
        display_name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        stripe_customer_id:
          type: ["string", "null"]
        approval_status:
          type: string
          enum: [pending, approved, rejected]
        created_by_user_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        member_role:
          type: string
          description: Only present on GET /businesses/mine responses
        member_status:
          type: string
          description: Only present on GET /businesses/mine responses
    BusinessInput:
      type: object
      required: [legal_name, display_name]
      properties:
        legal_name:
          type: string
        display_name:
          type: string
        address:
          type: string
        phone:
          type: string
    BusinessMemberInviteInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
        role:
          type: string
          enum: [owner, staff]
    BusinessMember:
      type: object
      required: [id, business_id, user_id, role, status, invited_at, updated_at]
      properties:
        id:
          type: integer
        business_id:
          type: integer
        user_id:
          type: integer
        role:
          type: string
          enum: [owner, staff]
        status:
          type: string
          enum: [active, pending, removed]
        spending_cap_cents:
          type: ["integer", "null"]
          description: Per-member spending cap in cents. NULL means no cap.
        invited_at:
          type: string
          format: date-time
        accepted_at:
          type: ["string", "null"]
          format: date-time
        updated_at:
          type: string
          format: date-time
        name:
          type: string
          description: Only present on GET /businesses/:id/members
        email:
          type: string
          description: Only present on GET /businesses/:id/members
        avatar_url:
          type: ["string", "null"]
          description: Only present on GET /businesses/:id/members
    BusinessMemberCapInput:
      type: object
      required: [spending_cap_cents]
      properties:
        spending_cap_cents:
          type: integer
          description: Per-member spending cap in cents. Use 0 to block all paid posts, or a positive value to cap total spend.
    BusinessRequest:
      type: object
      required: [id, title, status, payment_type, pay_it_forward_amount, requester_id, requester_name, created_at]
      properties:
        id:
          type: integer
        title:
          type: string
        status:
          type: string
        payment_type:
          type: string
        pay_it_forward_amount:
          type: ["number", "null"]
        requester_id:
          type: integer
        requester_name:
          type: string
        created_at:
          type: string
          format: date-time
    BusinessPendingRequest:
      type: object
      required: [id, title, status, payment_type, pay_it_forward_amount, requester_id, requester_name, created_at]
      properties:
        id:
          type: integer
        title:
          type: string
        status:
          type: string
        payment_type:
          type: string
        pay_it_forward_amount:
          type: ["number", "null"]
        requester_id:
          type: integer
        requester_name:
          type: string
        created_at:
          type: string
          format: date-time
    BusinessRequestApproveAction:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [approve, reject]
...